Move AnyKind and helper functions to RST stage
