+++ duo-lang-parser/src/Syntax/CST/Kinds.hs
@@ -49,7 +49,7 @@ data MonoKind
   deriving (Show, Eq, Ord)
 
 type MaybeKindedSkolem = (SkolemTVar, Maybe PolyKind)
-type KindedSkolem = (SkolemTVar,PolyKind)
+
 
 
 ------------------------------------------------------------------------------
@@ -65,41 +65,3 @@ data PolyKind =
 deriving instance (Show PolyKind)
 deriving instance (Eq PolyKind)
 deriving instance (Ord PolyKind)
-
---either polykind or primitive kind
-data AnyKind = MkPknd PolyKind | MkI64 | MkF64 | MkChar | MkString 
-deriving instance (Show AnyKind)
-deriving instance (Eq AnyKind)
-deriving instance (Ord AnyKind)
-
-monoToAnyKind :: MonoKind -> AnyKind
-monoToAnyKind (CBox eo) = MkPknd (MkPolyKind [] eo)
-monoToAnyKind I64Rep = MkI64
-monoToAnyKind F64Rep = MkF64
-monoToAnyKind CharRep = MkChar
-monoToAnyKind StringRep = MkString
-
-anyToMonoKind :: AnyKind -> MonoKind 
-anyToMonoKind (MkPknd (MkPolyKind _ eo)) = CBox eo
-anyToMonoKind MkI64 = I64Rep
-anyToMonoKind MkF64 = F64Rep
-anyToMonoKind MkChar = CharRep
-anyToMonoKind MkString = StringRep
-anyToMonoKind _ = error "should never happen"
-
-allTypeVars :: PolyKind -> Set SkolemTVar
-allTypeVars pk@MkPolyKind{} = S.fromList ((\(_,var,_) -> var) <$> pk.kindArgs)
-allTypeVars (KindVar _) = S.empty
-
-lookupPolyKind :: SkolemTVar -> PolyKind -> Maybe (Variance, SkolemTVar, MonoKind)
-lookupPolyKind tv pk@MkPolyKind{} = go pk.kindArgs
-  where
-    go :: [(Variance, SkolemTVar, MonoKind)] -> Maybe (Variance, SkolemTVar, MonoKind)
-    go [] = Nothing
-    go (k@(_,tv',_) : ks) = if tv == tv'
-                           then Just k
-                           else go ks
-lookupPolyKind _ (KindVar _) = Nothing
-
-lookupPolyKindVariance :: SkolemTVar -> PolyKind -> Maybe Variance
-lookupPolyKindVariance tv pk = (\(v,_,_) -> v) <$> lookupPolyKind tv pk
