+++ src/Dualize/Dualize.hs
@@ -13,6 +13,7 @@ import Syntax.RST.Terms (PrimitiveOp)
 import Syntax.RST.Types (FlipPol, FlipPrdCns, PolarityRep(..), flipPolarityRep, flipPrdCns)
 import Syntax.RST.Program (PrdCnsToPol)
 import Syntax.RST.Names
+import Syntax.RST.Kinds (AnyKind(..))
 import Syntax.Core.Terms (Pattern(..))
 import Syntax.TST.Types qualified as TST
 import Syntax.TST.Program qualified as TST
