+++ src/Pretty/Constraints.hs
@@ -10,6 +10,7 @@ import Syntax.TST.Types qualified as TST
 import Syntax.RST.Types (Polarity(..), PolarityRep (..))
 import Syntax.CST.Names
 import Syntax.RST.Names
+import Syntax.RST.Kinds
 import TypeInference.Constraints
 import Syntax.CST.Kinds
