+++ src/Translate/EmbedTST.hs
@@ -12,7 +12,8 @@ import Syntax.Core.Terms qualified as Core
 import Syntax.Core.Program qualified as Core
 
 import Data.Bifunctor (bimap, second)
-import Syntax.CST.Kinds (PolyKind(..), MonoKind(..), AnyKind(..))
+import Syntax.CST.Kinds (PolyKind(..), MonoKind(..))
+import Syntax.RST.Kinds (AnyKind(..))
 
 ---------------------------------------------------------------------------------
 -- A typeclass for embedding TST.X into Core.X
