+++ src/TypeAutomata/Definition.hs
@@ -15,7 +15,8 @@ import Syntax.CST.Names ( XtorName )
 import Syntax.CST.Types ( DataCodata(..), Arity, PrdCns(..))
 import Syntax.RST.Types ( Polarity, PolarityRep(..))
 import Syntax.RST.Names ( RnTypeName )
-import Syntax.CST.Kinds ( Variance, PolyKind(..),AnyKind(..))
+import Syntax.RST.Kinds
+import Syntax.CST.Kinds ( Variance, PolyKind(..))
 
 --------------------------------------------------------------------------------
 -- # Type Automata
