+++ src/TypeAutomata/FromAutomaton.hs
@@ -6,6 +6,7 @@ module TypeAutomata.FromAutomaton ( autToType ) where
 import Syntax.TST.Types
 import Syntax.RST.Types (PolarityRep(..), flipPolarityRep)
 import Syntax.RST.Names
+import Syntax.RST.Kinds
 import Syntax.CST.Types qualified as CST
 import Syntax.CST.Types (PrdCns(..), PrdCnsRep(..))
 import Syntax.CST.Names
