+++ src/TypeAutomata/ToAutomaton.hs
@@ -20,6 +20,7 @@ import Pretty.Pretty (ppPrint)
 import Syntax.TST.Types
 import Syntax.RST.Types (PolarityRep(..), Polarity(..), polarityRepToPol)
 import Syntax.RST.Names
+import Syntax.RST.Kinds
 import Syntax.CST.Types qualified as CST
 import Syntax.CST.Types (PrdCnsRep(..), PrdCns(..))
 import Syntax.CST.Names
