+++ src/TypeInference/Coalescing.hs
@@ -13,7 +13,7 @@ import Syntax.TST.Types
 import Syntax.RST.Types (PolarityRep(..),Polarity(..))
 import Syntax.CST.Names
 import Syntax.RST.Names
-import Syntax.CST.Kinds (AnyKind(..))
+import Syntax.RST.Kinds (AnyKind(..))
 import TypeInference.Constraints
 import Loc ( defaultLoc )
