+++ src/TypeInference/Constraints.hs
@@ -6,6 +6,7 @@ import Syntax.TST.Types
 import Syntax.CST.Names
 import Syntax.RST.Types (Polarity(..))
 import Syntax.RST.Names
+import Syntax.RST.Kinds
 import Syntax.CST.Kinds
 import Loc ( Loc )
