+++ src/TypeInference/GenerateConstraints/Definition.hs
@@ -55,6 +55,7 @@ import Syntax.CST.Types (PrdCnsRep(..), PrdCns(..))
 import Syntax.RST.Types (Polarity(..), PolarityRep(..))
 import Syntax.RST.Program as RST
 import Syntax.RST.Names
+import Syntax.RST.Kinds
 import Syntax.TST.Program as TST
 import Syntax.LocallyNameless (Index)
 import TypeInference.Constraints
