+++ src/TypeInference/GenerateConstraints/Kinds.hs
@@ -9,6 +9,7 @@ import Syntax.TST.Program qualified as TST
 import Syntax.RST.Program qualified as RST
 import Syntax.RST.Types qualified as RST
 import Syntax.RST.Names
+import Syntax.RST.Kinds
 import Syntax.TST.Types qualified as TST
 import Syntax.TST.Types (getKind)
 import Syntax.CST.Kinds
