+++ src/TypeInference/SolveConstraints.hs
@@ -24,6 +24,7 @@ import Errors
 import Syntax.TST.Types
 import Syntax.RST.Types (PolarityRep(..), Polarity(..))
 import Syntax.RST.Names
+import Syntax.RST.Kinds
 import Pretty.Pretty
 import Pretty.Types ()
 import Pretty.Constraints ()
